refactor: increase refinement quality
This commit enhances the quality of refinement in the resulting mesh.
The generated triangles are now not uniform as before, with smaller ones created
near boundaries and larger ones, not split, inside the bulk.
Additionally, some parameters have been marked as obsolete, and a new
parameter, `ConcentricShellParameter`, has been introduced in the
triangulator settings. The `RefineJob` contract has been simplified and
is no longer a generic job.