Add ability to ignore some positions
When creating a new triangulation job, you can now optionally pass a NativeArray<bool> called IgnorePositions. This array should be the same length as the Positions array. Any indexes that are true it IngorePositions will not be used in the triangulation.

Example use case: regenerate a mesh's triangles but only use a subset of its vertices, without having to change the mesh's vertices array.